rfm69-async: Split lib.rs into multiple files
Also fix some minor clippy issues.
